Dockerfile: use newer busybox version
Image versions using in testing, such as busybox, are frozen to ensure
reproducibility. The version of busybox used upstream is old and doesn't
use a separate tmpfs for /run, instead symlinking /run to /tmp. This was
changed afterwards[1].

This then caused the mount to change from /run/secrets mount to
/tmp/secrets during container initialization since symlinks are resolved
(see the call to FollowSymlinkInScope inside rootfs_linux.go). This
change of mount in turn causes test failures which sometimes use /tmp as
a read-only mount.

This patch makes the testing platform use a more recent version of
busybox, where run is a separate tmpfs. It also updates the
TestDiffEnsureOnlyKmsgAndPtmx test to expect /run and /run/secrets.

Resolves: RHBZ1271334

[1] http://git.buildroot.net/buildroot/commit/?id=d380dde1d2878

Signed-off-by: Jonathan Lebon <jlebon@redhat.com>