Don't try to look up the name for unset login UIDs
When logging the client request to syslog, if the login UID of the
client process is not set (i.e., reading it gives us the value -1),
don't bother trying to look up the name of the user associated with that
UID.  This is the case for services started by init.  Fixes
RHBZ#1275399.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com> (github:nalind)