Send logs to the journal "from" the container
Previously, dockerd just relayed messages by itself from containers to
the journal, which caused journald to apply rate limiting to messages
across all containers as a single group.

Here, we add another process to each container's cgroup, and have
dockerd forward messages to that process over a pipe.  That process,
named "journal-logger", receives the messages and sends them on to the
journal.

As part of the container's cgroup, it's killed when the main process
exits, so we only need to close the pipe and read its exit status when
closing the logger.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>