Store signatures in c/i/docker/daemon/signatures
Whether or not we are verifying signatures, download them and store them
in docker/daemon/signatures.

Note that this means that containers/image/docker is now involved on
_every_ pull; failures of the c/i/docker client, or inability to
download (possibly incorrectly configured but unused) signatures are now
fatal.

Alternatively, we could make the storing of signatures to c/i/d/d/s silently
fail in such cases.

WARNING: This DOES NOT BUILD because it references sirupsen/logrus, not
Sirupsen/logrus.