allow storing images with >= 127 layers
There is an overlay2 restriction that limits the maximum number of
(lower) layers to 127 which is induced by kernel's maximum arguments and
possible the size as well when mounting.  The checks are very strict in
that we cannot even store images exceeding this limit.  However, some
images might not be meant to be running in a container.  Hence, defer
checking the number of (lower) layers right before mounting trying to
mount a layer in overlay2.  Also remove the #layer checks when storing
images.  Now, we can store but not mount these images.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>